Additional game components added. Refined test cases. Component collection types added. Redefined datastore interface. SQlite datastore is currently broken, pending implementation of new datastore interface.

+++ data/template/store.go
@@ -16,7 +16,7 @@ func (this *Store) Initialize(*builder.World) (err os.Error)     { return }
 func (this *Store) GetWorld() (world lib.World, err os.Error) { return }
 func (this *Store) SetWorld(world *lib.World) (err os.Error)  { return }
 
-func (this *Store) GetUser(id uint16) (usr *lib.UserInfo, err os.Error)         { return }
+func (this *Store) GetUser(id int64) (usr *lib.UserInfo, err os.Error)          { return }
 func (this *Store) GetUserByName(name string) (usr *lib.UserInfo, err os.Error) { return }
 func (this *Store) SetUser(usr *lib.UserInfo) (err os.Error)                    { return }
 func (this *Store) GetUsers() (usr []*lib.UserInfo, err os.Error)               { return }
